# services/identity/openapi/identity.yaml
openapi: 3.1.0
info:
  title: Identity Service
  version: 1.0.0
  description: >
    Staff identity, sessions, roles and permissions. Every /api/v1 route except
    login requires a signed principal, forwarded by the gateway after it
    resolves the session cookie. Services never call this API to authenticate a
    request; they verify the principal's signature.

    Identifiers are strings on every surface. They are BIGSERIAL values that can
    exceed JavaScript's safe integer range, where a JSON number would silently
    lose precision.
servers:
  - url: http://localhost:8080
    description: Through the gateway
paths:
...
  /api/v1/logout:
    post:
      summary: Revoke the caller's session.
      responses:
        "204": { description: Revoked }
        "401": { $ref: "#/components/responses/Error" }
...
  /api/v1/me/password:
    post:
      summary: Change the caller's own password.
      description: >
        Reachable while mustChangePassword is set; every other route is refused
        until it clears.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordRequest" }
      responses:
        "204": { description: Changed }
        "401": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
components:
  responses:
    Error:
      description: Standard error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - invalid_credentials
            - unauthenticated
            - forbidden
            - password_change_required
            - not_found
            - conflict
            - last_admin
            - validation_failed
            - internal
            - schema_dirty
            - schema_not_migrated
            - database_unavailable
        message: { type: string }
...
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string }
...